LayoutTests/platform/glib/imported/w3c/web-platform-tests/clipboard-apis/async-navigator-clipboard-basics.https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://web-platform.test:9443/clipboard-apis/async-navigator-clipboard-basics.https.html was not allowed to display insecure content from http://localhost:8001/clipboard-apis/resources/greenbox.png.

CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8001/clipboard-apis/resources/greenbox.png due to access control checks.

PASS navigator.clipboard exists
PASS navigator.clipboard.write([text/plain ClipboardItem]) succeeds
PASS navigator.clipboard.write() fails (expect [ClipboardItem])
PASS navigator.clipboard.write(null) fails (expect [ClipboardItem])
PASS navigator.clipboard.write(DOMString) fails (expect [ClipboardItem])
PASS navigator.clipboard.write(Blob) fails (expect [ClipboardItem])
PASS navigator.clipboard.writeText(DOMString) succeeds
PASS navigator.clipboard.writeText() fails (expect DOMString)
FAIL navigator.clipboard.write({string : image/png Blob}) succeeds promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL navigator.clipboard.read() succeeds promise_test: Unhandled rejection with value: object "NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission."
FAIL navigator.clipboard.readText() succeeds promise_test: Unhandled rejection with value: object "NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission."
